requirements.txt
langchain>=0.0.335
langchain-openai>=0.0.6
langchain-community>=0.0.13
langgraph>=0.0.20
python-dotenv>=1.0.0
serpapi>=0.1.0
openai>=1.6.0
langchain-experimental>=0.0.40
